Alpha implementation of OPML import with folders specified in OPML file
To clarify, the goal is to have in my OPML file containing various RSS sources, that should go into various IMAP folders. To ensure it is easy, OPML file is described the following way

    <outline text="work" title="RSS/work">
         <outline text="The Tao of Mac" type="rss" title="The Tao of Mac" xmlUrl="http://taoofmac.com/rss" htmlUrl="http://taoofmac.com" />
         .
         .
         .
     </outline>

This will import all RSS feeds into the RSS/work folder, whatever the default RSS folder is.
However, to ensure this behaviour stays optional, a config.py option has been added : USE_OPML_TITLE_AS_FOLDER, which must be set to true to enable this feature. I'm not sure it's the best way, but it ensure all this behaviour is an opt-in option.